Starting to get messy!
Added code to generate a tree of all stock. It's really nice
Will have to see how big of a json that is. Maybe better to do on phone.

Also started to make read messy.
Need to add an "all" parameter that sums it up by store_id.
Maybe it can do that first and then we don't have to worry...